app/worker/requirements.txt
SQLAlchemy
python-decouple
fastapi
uvicorn
passlib
pydantic
numpy
joblib
pydantic-settings
psycopg2-binary
asyncpg
scikit-learn
python-dotenv
passlib
pika
pydantic
alembic
psycopg
PyJWT
python-multipart
streamlit
requests